Ensure request attributes are propegated into async requests as well
